img_classifier/app/requirements.txt
tensorflow_hub==0.12.0
pillow==9.0.1
scikit-learn==1.0.1
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
faiss-cpu==1.7.1.post2